requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
plotly==5.17.0
scipy==1.11.4

# NLP and ML
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
nltk==3.8.1

# Data processing
openpyxl==3.1.2
python-docx==1.1.0
Jinja2==3.1.2

# APIs and integrations
gspread==5.12.0
google-auth==2.23.4
openai==1.3.7
requests==2.31.0

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Database
sqlite3  # Built-in with Python

# Optional: For advanced NLP in Portuguese
# sentence-transformers==2.2.2